got rid of separate classes for different datasets
