+++ config.py
@@ -2,6 +2,9 @@
 
 import argparse
 
+from layers.graph_layer.edge_functions import edge_f_dict as d
+
+
 def get_parser(name):
     """
 
@@ -12,7 +15,7 @@ def get_parser(name):
     parser = argparse.ArgumentParser(name, formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     return parser
 
-class Config(argparse.Namespace):
+class MNISTConfig(argparse.Namespace):
     def build_parser(self):
         parser = get_parser("InvNet config")
         parser.add_argument('--dataset', default='mnist', help='circle / polycrystalline')
@@ -20,6 +23,7 @@ class Config(argparse.Namespace):
         parser.add_argument('--output_path', default='./output_dir', help='output directory')
         parser.add_argument('--data_dir', default='/data/MNIST')
         parser.add_argument('--gpu', default=1, help='Selecting the gpu')
+        parser.add_argument('--data_size', default=64, type=int)
         parser.add_argument('--batch_size', default=32,type=int, help='Batch size for training')
         parser.add_argument('--hidden_size', default=32, type=int,help='Hidden size used for generator and discriminator')
         parser.add_argument('--critic_iter', default=5, type=int,help='Number of iter for descriminator')
@@ -29,10 +33,9 @@ class Config(argparse.Namespace):
         parser.add_argument('--restore_mode', default=False,
                             help='If True, it will load saved model from OUT_PATH and continue to train')
 
-        parser.add_argument('--max_op', default=False,
-                            help='If True, it will load saved model from OUT_PATH and continue to train')
-        parser.add_argument('--edge_fn', default='diff_exp',
-                            help='If True, it will load saved model from OUT_PATH and continue to train')
+        parser.add_argument('--max_op', default=False)
+        parser.add_argument('--edge_fn', default='diff_exp')
+        parser.add_argument('--make_pos', type=bool,default=True)
         return parser
 
     def __init__(self):
@@ -40,27 +43,27 @@ class Config(argparse.Namespace):
         args = parser.parse_args()
         super().__init__(**vars(args))
 
-class TestConfig(argparse.Namespace):
+
+class MicroStructureConfig(argparse.Namespace):
     def build_parser(self):
-        parser = get_parser("InvNet config")
-        parser.add_argument('--dataset', default='mnist', help='circle / polycrystalline')
+        parser = get_parser("MicroConfig config")
         parser.add_argument('--lr',default=01e-04)
         parser.add_argument('--output_path', default='./output_dir', help='output directory')
-        parser.add_argument('--data_dir', default='/data/MNIST')
+        parser.add_argument('--data_dir', default='/data/datasets/two_phase_morph/')
         parser.add_argument('--gpu', default=1, help='Selecting the gpu')
-        parser.add_argument('--batch_size', default=4,type=int, help='Batch size for training')
-        parser.add_argument('--hidden_size', default=4, type=int,help='Hidden size used for generator and discriminator')
-        parser.add_argument('--critic_iter', default=1, type=int,help='Number of iter for descriminator')
-        parser.add_argument('--proj_iter', default=3, type=int, help='Number of iteration for projection update.')
-        parser.add_argument('--end_iter', default=5000, help='How many iterations to train for.')
+        parser.add_argument('--data_size',default=64,type=int)
+        parser.add_argument('--batch_size', default=32,type=int, help='Batch size for training')
+        parser.add_argument('--hidden_size', default=32, type=int,help='Hidden size used for generator and discriminator')
+        parser.add_argument('--critic_iter', default=5, type=int,help='Number of iter for descriminator')
+        parser.add_argument('--proj_iter', default=1, type=int, help='Number of iteration for projection update.')
+        parser.add_argument('--end_iter', default=20000, help='How many iterations to train for.')
         parser.add_argument('--lambda_gp', default=10, help='gradient penalty hyperparameter')
         parser.add_argument('--restore_mode', default=False,
                             help='If True, it will load saved model from OUT_PATH and continue to train')
 
-        parser.add_argument('--max_op', default=False,
-                            help='If True, it will load saved model from OUT_PATH and continue to train')
-        parser.add_argument('--edge_fn', default='diff_exp',
-                            help='If True, it will load saved model from OUT_PATH and continue to train')
+        parser.add_argument('--max_op', default=False)
+        parser.add_argument('--edge_fn', choices=list(d.keys()),default='diff_exp')
+        parser.add_argument('--make_pos', type=bool, default=False)
         return parser
 
     def __init__(self):
