+++ graph_invnet/__init__.py
@@ -1,2 +1,2 @@
-from .invnet import BaseInvNet
+from .invnet import GraphInvNet
 from .utils import calc_gradient_penalty, weights_init
