+++ graph_invnet/utils.py
@@ -1,7 +1,10 @@
+import torch.nn as nn
+import h5py
 import torch
 import torch.nn as nn
 import torch.nn.init as init
 from torch import autograd
+from torch.utils.data import Dataset
 
 from models.wgan import MyConvo2d
 
@@ -47,5 +50,17 @@ def calc_gradient_penalty(netD, real_data, fake_data,batch_size,lambd,size):
     gradient_penalty = ((gradients.norm(2, dim=1) - 1) ** 2).mean() * lambd
     return gradient_penalty
 
+class MicrostructureDataset(Dataset):
+    def __init__(self, data_path, transform=None):
+        super(MicrostructureDataset, self).__init__()
+        self.data = h5py.File(data_path, mode='r')['morphology_64_64']
+        self.transform = transform
 
+    def __getitem__(self, index):
+        x = torch.FloatTensor(self.data[index, ...])
+        if self.transform is not None:
+            x = self.transform(x)
+        return x/255
 
+    def __len__(self):
+        return self.data.shape[0]
