+++ layers/__init__.py
@@ -1,2 +1,2 @@
 from layers.dp_layer.DPLayer import DPLayer
-from layers.dp_layer.edge_functions import edge_f_dict
+from layers.graph_layer.edge_functions import edge_f_dict
