+++ layers/tests/test_graph_layer.py
@@ -2,8 +2,8 @@ import numpy as np
 import torch
 from scipy.optimize import check_grad
 
-from layers.dp_layer.edge_functions import sum_squared
 from layers.graph_layer import GraphLayer
+from layers.graph_layer.edge_functions import sum_squared
 
 
 def make_data():
