--- microstructure/__init__.py
@@ -1,3 +1,0 @@
-from .config import Config
-from .micro_invnet import InvNet
-from .utils import MicrostructureDataset
