--- mnist/__init__.py
@@ -1,2 +1,0 @@
-from .config import Config,TestConfig
-from .mnist_invnet import InvNet
